+++ osf/management/commands/migratedata.py
@@ -7,6 +7,7 @@ import pstats
 import sys
 from cProfile import Profile
 
+import bson
 import ipdb
 from django.contrib.contenttypes.models import ContentType
 from django.core.management import BaseCommand
@@ -24,8 +25,6 @@ from framework.mongo import database
 from framework.mongo import set_up_storage
 from framework.mongo import storage
 from framework.transactions.context import transaction as modm_transaction
-from osf.management.commands.migraterelations import find_duplicate_addon_node_settings
-from osf.management.commands.migraterelations import find_duplicate_addon_user_settings
 from osf.models import Comment
 from osf.models import (NodeLog, OSFUser,
                         PageCounter, StoredFileNode, Tag, UserActivityCounter)
@@ -742,6 +741,138 @@ def merge_duplicate_users():
     logger.info('Done with {} in {} seconds...'.format(
         sys._getframe().f_code.co_name,
         (timezone.now() - start).total_seconds()))
+def find_duplicate_addon_user_settings():
+    COLLECTIONS = [
+        'addondataverseusersettings',
+        'addonfigshareusersettings',
+        # 'addongithubusersettings',  # old, unused
+        'addonowncloudusersettings',
+        # 'addons3usersettings',  # old, unused
+        'boxusersettings',
+        'dropboxusersettings',
+        'githubusersettings',
+        'googledriveusersettings',
+        'mendeleyusersettings',
+        'osfstorageusersettings',
+        's3usersettings',
+        'zoterousersettings',
+        'addonwikiusersettings'
+    ]
+    from framework.mongo.handlers import database
+    for collection in COLLECTIONS:
+        targets = database[collection].aggregate([
+            {
+                "$group": {
+                    "_id": "$owner",
+                    "ids": {"$addToSet": "$_id"},
+                    "count": {"$sum": 1}
+                }
+            },
+                {
+                    "$match": {
+                        "count": {"$gt": 1}
+                    }
+                },
+                {
+                    "$sort": {
+                        "count": -1
+                    }
+                }
+        ]).get('result')
+        for group in targets:
+            oauth_grants = {}
+            bad = []
+            good = []
+            newest = None
+            for _id in group['ids']:
+                instance = database[collection].find_one(_id)
+                if 'oauth_grants' in instance:
+                    oauth_grants.update(instance['oauth_grants'])
+                if instance['deleted']:
+                    bad.append(instance)
+                else:
+                    good.append(instance)
+            for us in good:
+                if not newest or bson.ObjectId(us['_id']).generation_time > bson.ObjectId(newest['_id']).generation_time:
+                    newest = us
+            # remove the keeper
+            good.pop(good.index(newest))
+            logger.info('Updating {} oauth_grants to {}'.format(newest['_id'], oauth_grants))
+            database[collection].update({'_id': newest['_id']}, {"$set": {'oauth_grants': oauth_grants}})
+            for worst in good + bad:
+                logger.info('Removing {} from {}'.format(worst['_id'], collection))
+                database[collection].remove(worst['_id'])
+
+
+def find_duplicate_addon_node_settings():
+    COLLECTIONS = [
+        'addondataversenodesettings',
+        'addonfigsharenodesettings',
+        # 'addongithubnodesettings',  # old, unused
+        'addonowncloudnodesettings',
+        # 'addons3nodesettings',  # old, unused
+        'boxnodesettings',
+        'dropboxnodesettings',
+        'githubnodesettings',
+        'googledrivenodesettings',
+        'mendeleynodesettings',
+        'osfstoragenodesettings',
+        's3nodesettings',
+        'zoteronodesettings',
+        'addonwikinodesettings'
+    ]
+    from framework.mongo.handlers import database
+
+    for collection in COLLECTIONS:
+        targets = database[collection].aggregate([
+            {
+                "$group": {
+                    "_id": "$owner",
+                    "ids": {"$addToSet": "$_id"},
+                    "count": {"$sum": 1}
+                }
+            },
+            {
+                "$match": {
+                    "count": {"$gt": 1}
+                }
+            },
+            {
+                "$sort": {
+                    "count": -1
+                }
+            }
+        ]).get('result')
+        for group in targets:
+            if not group['_id']:
+                for _id in group['ids']:
+                    if _id:
+                        logger.info('Removing {} from {}'.format(_id, collection))
+                        database[collection].remove(_id)
+                    else:
+                        logger.info('_id was None: {}'.format(group))
+            else:
+                active_ns = []
+                for _id in group['ids']:
+                    if database[collection].find_one(_id)['external_account']:  # or w/e key it is
+                        active_ns.append(_id)
+                if not active_ns:
+                    good = group['ids'].pop()
+                    for bad in group['ids']:
+                        if bad:
+                            logger.info('Removing {} from {}'.format(bad, collection))
+                            database[collection].remove(bad)
+                        else:
+                            logger.info('_id was None: {}'.format(group))
+                elif len(active_ns) == 1:
+                    for _id in group['ids']:
+                        if _id  and _id not in active_ns:
+                            logger.info('Removing {} from {}'.format(_id, collection))
+                            database[collection].remove(_id)
+                        else:
+                            logger.info('_id was None: {}'.format(group))
+                else:
+                    raise Exception('Something untoward happened.')
 
 
 class Command(BaseCommand):
